rename:충돌 방지 auth 이름 변경
