+++ src/(pages)/auth/login/LoginPage.tsx
@@ -1,4 +1,4 @@
-import { AuthLayout } from '@/features/auth/ui';
+import { AuthLayout } from '@/features/auth-login/ui';
 import { LoginForm } from '@/widgets/login/ui';
 export default function LoginPage() {
   return (
